requirements.txt
discord.py==2.1.1
pygsheet==0.1.19b25
python-dotenv==0.21.1
